Allow mgo-statsd to be called as a module
